requirements.txt
opencv-python
kornia>=0.6.11
Pillow>=9.2.0
scipy>=1.0.0
yacs
matplotlib==3.4.2
wandb>=0.15.4
pytorch-lightning>=2.0.0
pre-commit>=2.0
isort
flake8==4.0.1
pydocstyle
black
mypy
rtpt==0.0.4
timm==0.9.12
pycocotools
ipython
scikit-learn==1.3.2
pykeops>=2.1.2
av>=11.0.0
torchmetrics==1.3.1
cython
scikit-image>=0.22.0
optuna
colored
faster-coco-eval
open3d
einops
